+++ tests/ByteDev.Crypto.UnitTests/StringExtensions.cs
@@ -19,5 +19,20 @@ namespace ByteDev.Crypto.UnitTests
 
             return true;
         }
+
+        public static bool IsHex(this string source)
+        {
+            foreach(char c in source)
+            {
+                var isHex = c >= '0' && c <= '9' || 
+                            c >= 'A' && c <= 'F' ||
+                            c >= 'a' && c <= 'f';
+
+                if(!isHex)
+                    return false;
+            }
+
+            return true;
+        }
     }
 }